improve reload arch, reduce redundancy in ping, jitter

+++ src/lib/utils/types.go
@@ -61,9 +61,9 @@ type FloodPingResp struct {
 // This is the object that we return from resp_delay module
 // Contains delay in monitor and the monitor length
 type Response struct {
-	Delay         float64 `json:"delay"`
-	ResLength     int     `json:"resLength"`
-	ResStatusCode int     `json:"resStatusCode"`
+	Delay         int64 `json:"delay"`
+	ResLength     int   `json:"resLength"`
+	ResStatusCode int   `json:"resStatusCode"`
 }
 
 // ResponseResp for responding the querier.
